refactor: remove old framer plugin
