--- packages/obra-icons-framer-plugin-new/framer.json
@@ -1,7 +1,0 @@
-{
-    "id": "7aad7a",
-    "name": "Obra Icons",
-    "description": "Add beautiful, customizable icons to your Framer projects",
-    "modes": ["canvas"],
-    "icon": "/icon.png"
-}
