--- packages/obra-icons-framer-plugin-new/index.html
@@ -1,15 +1,0 @@
-<!doctype html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <link rel="icon" type="image/svg+xml" href="/framer.svg" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>Phosphor Icons</title>
-        <link rel="preconnect" href="https://fonts.googleapis.com" />
-    </head>
-
-    <body>
-        <div id="root"></div>
-        <script type="module" src="/src/main.tsx"></script>
-    </body>
-</html>
