--- packages/obra-icons-framer-plugin-new/package.json
@@ -1,33 +1,0 @@
-{
-    "name": "obra-icons-framer-plugin",
-    "private": true,
-    "version": "0.0.0",
-    "type": "module",
-    "main": "src/api/index.ts",
-    "scripts": {
-        "dev": "vite",
-        "build": "vite build",
-        "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-        "preview": "vite preview",
-        "pack": "npx framer-plugin-tools@latest pack",
-        "check-biome": "run g:check-biome",
-        "check-eslint": "run g:check-eslint",
-        "check-typescript": "run g:check-typescript"
-    },
-    "dependencies": {
-        "obra-icons-react": "^1.23.1",
-        "framer-plugin": "^3.3.2",
-        "fuse.js": "^7.1.0",
-        "react": "^18.3.1",
-        "react-dom": "^18.3.1",
-        "obra-icons-website": "workspace:*"
-    },
-    "devDependencies": {
-        "@types/react": "^18.3.23",
-        "@types/react-dom": "^18.3.7",
-        "@vitejs/plugin-react-swc": "^3.11.0",
-        "vite": "^7.0.5",
-        "vite-plugin-framer": "^1.0.7",
-        "vite-plugin-mkcert": "^1.17.8"
-    }
-}
