--- packages/obra-icons-framer-plugin-new/src/vite-env.d.ts
@@ -1,5 +1,0 @@
-/// <reference types="vite/client" />
-
-interface ViteTypeOptions {
-    strictImportMetaEnv: unknown
-}
