--- packages/obra-icons-framer-plugin/.gitignore
@@ -1,33 +1,0 @@
-# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
-
-# dependencies
-
-/node_modules
-/.pnp
-.pnp.js
-.yarn
-
-# misc
-
-.DS_Store
-\*.pem
-
-# files
-
-my-plugin
-dev-plugin
-dist
-
-# debug
-
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log\*
-
-# local env files
-
-.env\*.local
-
-# Packed plugin
-
-plugin.zip
