+++ packages/obra-icons-framer-plugin/framer.json
@@ -1,9 +1,7 @@
 {
-  "id": "5ec1f9",
-  "name": "Obra Icons",
-  "description": "Add beautiful, customizable icons to your Framer projects",
-  "modes": [
-    "canvas"
-  ],
-  "icon": "/icon.svg"
+    "id": "7aad7a",
+    "name": "Obra Icons",
+    "description": "Add beautiful, customizable icons to your Framer projects",
+    "modes": ["canvas"],
+    "icon": "/icon.png"
 }
