+++ packages/obra-icons-framer-plugin/index.html
@@ -1,14 +1,15 @@
-<!DOCTYPE html>
-  <html lang="en">
+<!doctype html>
+<html lang="en">
     <head>
-      <meta charset="UTF-8" />
-      <link rel="icon" type="image/svg+xml" href="/icon.svg" />
-      <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-      <title>Obra Icons Framer Plugin</title>
+        <meta charset="UTF-8" />
+        <link rel="icon" type="image/svg+xml" href="/framer.svg" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <title>Phosphor Icons</title>
+        <link rel="preconnect" href="https://fonts.googleapis.com" />
     </head>
+
     <body>
-      <div id="root"></div>
-      <script type="module" src="/src/main.tsx"></script>
+        <div id="root"></div>
+        <script type="module" src="/src/main.tsx"></script>
     </body>
-  </html>
-  
+</html>
