+++ packages/obra-icons-framer-plugin/package.json
@@ -1,37 +1,33 @@
 {
-  "name": "obra-icons-framer-plugin",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-    "preview": "vite preview",
-    "pack": "npx framer-plugin-tools@latest pack"
-  },
-  "dependencies": {
-    "framer-plugin": "^3",
-    "react": "^18",
-    "react-dom": "^18",
-    "vite-plugin-mkcert": "^1",
-    "@orama/orama": "=1.2.8",
-    "obra-icons-react": "workspace:*",
-    "obra-icons-website": "workspace:*"
-  },
-  "devDependencies": {
-    "@eslint/js": "^9",
-    "@types/react-dom": "^18",
-    "@types/react": "^18",
-    "@vitejs/plugin-react-swc": "^3",
-    "@vitejs/plugin-react": "^4.3.1",
-    "eslint-plugin-react-hooks": "^5.1.0",
-    "eslint-plugin-react-refresh": "^0.4.9",
-    "typescript-eslint": "^8.0.1",
-    "vite-plugin-framer": "^1",
-    "eslint": "^9.9.0",
-    "globals": "^15.9.0",
-    "typescript": "^5.3",
-    "vite": "^6"
-  }
+    "name": "obra-icons-framer-plugin",
+    "private": true,
+    "version": "0.0.0",
+    "type": "module",
+    "main": "src/api/index.ts",
+    "scripts": {
+        "dev": "vite",
+        "build": "vite build",
+        "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+        "preview": "vite preview",
+        "pack": "npx framer-plugin-tools@latest pack",
+        "check-biome": "run g:check-biome",
+        "check-eslint": "run g:check-eslint",
+        "check-typescript": "run g:check-typescript"
+    },
+    "dependencies": {
+        "obra-icons-react": "^1.23.1",
+        "framer-plugin": "^3.3.2",
+        "fuse.js": "^7.1.0",
+        "react": "^18.3.1",
+        "react-dom": "^18.3.1",
+        "obra-icons-website": "workspace:*"
+    },
+    "devDependencies": {
+        "@types/react": "^18.3.23",
+        "@types/react-dom": "^18.3.7",
+        "@vitejs/plugin-react-swc": "^3.11.0",
+        "vite": "^7.0.5",
+        "vite-plugin-framer": "^1.0.7",
+        "vite-plugin-mkcert": "^1.17.8"
+    }
 }
