+++ packages/obra-icons-framer-plugin/public/icon.png
