+++ packages/obra-icons-framer-plugin/src/App.css
@@ -1,12 +1,95 @@
-/* Obra Icons Framer Plugin CSS */
+#root {
+	width: 100%;
+	display: flex;
+	flex-direction: column;
+	padding: 0 15px 0px 15px;
+	gap: 10px;
+}
 
-.obra-icons-plugin {
-    padding: 16px;
-    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
-    font-size: 14px;
-    height: 100%;
-    background-color: red;
-    display: flex;
-    flex-direction: column;
-  }
-    
+.weight-selector {
+	width: 100%;
+}
+
+.type-selector {
+	width: 100%;
+}
+
+.error-container {
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	flex-direction: column;
+	height: 100%;
+	padding: 0 15px 25px 15px;
+}
+
+.error-container p {
+	color: var(--framer-color-text-tertiary);
+}
+
+.search-container {
+	display: flex;
+	flex-direction: column;
+	gap: 10px;
+}
+
+.grid-container {
+	height: 100%;
+	width: 100%;
+	border-radius: 8px 8px 0 0;
+	overflow: hidden;
+	overflow-y: scroll;
+}
+
+.search-input {
+	height: 30px;
+	width: 100%;
+	font-size: 12px;
+	line-height: 24px;
+	font-weight: 500 !important;
+	font-family: 'Inter', Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
+}
+
+.search-input::placeholder {
+	outline: none;
+	border: none;
+	color: var(--framer-color-text-tertiary);
+	font-weight: 500;
+	font-size: 12px;
+	font-family: 'Inter', Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
+}
+
+.grid {
+	width: 100%;
+	display: grid;
+	border-radius: 8px;
+	overflow: hidden;
+	grid-template-columns: repeat(auto-fill, minmax(60px, 1fr));
+	grid-auto-rows: minmax(0px, 1fr);
+	padding-bottom: 15px;
+	gap: 10px;
+}
+
+.icon-parent {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	height: unset;
+	background: var(--framer-color-bg-tertiary);
+	aspect-ratio: 1 / 1;
+	border-radius: 8px;
+	cursor: pointer;
+
+	&:hover,
+	&:focus {
+		background: var(--framer-color-bg-tertiary);
+	}
+}
+
+::-webkit-scrollbar {
+	width: 0px;
+	height: 0px;
+}
+::-webkit-scrollbar-thumb {
+	background: transparent;
+}
