--- packages/obra-icons-framer-plugin/src/side/App.css
@@ -1,271 +1,0 @@
-/* Obra Icons Framer Plugin CSS */
-
-.obra-icons-plugin {
-  padding: 16px;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
-  font-size: 14px;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-}
-
-.toolbar {
-  margin-bottom: 16px;
-  border-bottom: 1px solid #e0e0e0;
-  padding-bottom: 16px;
-}
-
-.controls {
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-}
-
-.control-group {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-}
-
-.control-group label {
-  font-weight: 500;
-  font-size: 12px;
-  color: #666;
-}
-
-.control-group input[type="range"] {
-  width: 100%;
-}
-
-.control-group span {
-  font-size: 12px;
-  color: #888;
-}
-
-.fieldset {
-  border: none;
-  padding: 0;
-  margin: 0;
-}
-
-.legend {
-  font-weight: 500;
-  font-size: 12px;
-  color: #666;
-  margin-bottom: 8px;
-}
-
-.radio-buttons {
-  display: flex;
-  gap: 8px;
-}
-
-.radio-buttons label {
-  display: flex;
-  align-items: center;
-  gap: 4px;
-  font-size: 12px;
-  cursor: pointer;
-}
-
-.color-picker {
-  position: relative;
-}
-
-.selected-color {
-  width: 32px;
-  height: 32px;
-  border: 2px solid #ddd;
-  border-radius: 4px;
-  cursor: pointer;
-  background: white;
-}
-
-.menu-overlay {
-  position: absolute;
-  top: 100%;
-  left: 0;
-  background: white;
-  border: 1px solid #ddd;
-  border-radius: 8px;
-  padding: 12px;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
-  z-index: 1000;
-  min-width: 200px;
-}
-
-.color-grid {
-  display: grid;
-  grid-template-columns: repeat(5, 1fr);
-  gap: 4px;
-  margin-bottom: 8px;
-}
-
-.color-option {
-  width: 24px;
-  height: 24px;
-  border: 1px solid #ddd;
-  border-radius: 4px;
-  cursor: pointer;
-  padding: 0;
-}
-
-.add-custom-color {
-  background: #f5f5f5;
-  border: 1px solid #ddd;
-  border-radius: 4px;
-  padding: 8px;
-  cursor: pointer;
-  font-size: 16px;
-}
-
-.overlay-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 12px;
-}
-
-.overlay-header h3 {
-  margin: 0;
-  font-size: 14px;
-}
-
-.icon-button {
-  background: none;
-  border: 1px solid #ddd;
-  border-radius: 4px;
-  padding: 6px 8px;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  gap: 4px;
-  font-size: 12px;
-}
-
-.icon-button:hover {
-  background: #f5f5f5;
-}
-
-.add-color {
-  display: flex;
-  gap: 8px;
-  align-items: center;
-  margin-bottom: 12px;
-}
-
-.color-item {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  padding: 4px 0;
-}
-
-.color-sample {
-  width: 16px;
-  height: 16px;
-  border: 1px solid #ddd;
-  border-radius: 2px;
-}
-
-.search-holder {
-  position: relative;
-  margin-bottom: 16px;
-}
-
-.search-holder input {
-  width: 100%;
-  padding: 8px 32px 8px 12px;
-  border: 1px solid #ddd;
-  border-radius: 6px;
-  font-size: 14px;
-  box-sizing: border-box;
-}
-
-.clear-search {
-  position: absolute;
-  right: 8px;
-  top: 50%;
-  transform: translateY(-50%);
-  background: none;
-  border: none;
-  font-size: 16px;
-  cursor: pointer;
-  color: #666;
-  padding: 4px;
-}
-
-.clear-search:hover {
-  color: #000;
-}
-
-.loading {
-  text-align: center;
-  padding: 24px;
-  color: #666;
-  font-size: 14px;
-}
-
-.icon-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(48px, 1fr));
-  gap: 8px;
-  padding: 8px 0;
-  flex: 1;
-  overflow-y: auto;
-  max-height: 400px;
-}
-
-.icon-item {
-  background: white;
-  border: 1px solid #e0e0e0;
-  border-radius: 6px;
-  padding: 12px;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  min-height: 48px;
-  transition: all 0.2s ease;
-}
-
-.icon-item:hover {
-  border-color: #007AFF;
-  background: #f8f9ff;
-  transform: translateY(-1px);
-  box-shadow: 0 2px 8px rgba(0, 122, 255, 0.1);
-}
-
-.icon-item:active {
-  transform: translateY(0);
-}
-
-.no-results {
-  text-align: center;
-  padding: 24px;
-  color: #666;
-  font-size: 14px;
-  grid-column: 1 / -1;
-}
-
-.framer-button-primary {
-  background: #007AFF;
-  color: white;
-  border: none;
-  border-radius: 6px;
-  padding: 12px 24px;
-  font-size: 14px;
-  font-weight: 500;
-  cursor: pointer;
-  margin: 12px 0;
-}
-
-.framer-button-primary:hover {
-  background: #0056CC;
-}
-
-.error {
-  color: #FF3B30;
-  font-size: 12px;
-  margin-top: 4px;
-}
