+++ packages/obra-icons-framer-plugin/src/vite-env.d.ts
@@ -1,1 +1,5 @@
 /// <reference types="vite/client" />
+
+interface ViteTypeOptions {
+    strictImportMetaEnv: unknown
+}
