+++ packages/obra-icons-framer-plugin/tsconfig.base.json
