+++ packages/obra-icons-framer-plugin/vite.config.ts
@@ -1,12 +1,12 @@
-import { defineConfig } from "vite"
-import react from "@vitejs/plugin-react-swc"
-import mkcert from "vite-plugin-mkcert"
-import framer from "vite-plugin-framer"
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react-swc";
+import mkcert from "vite-plugin-mkcert";
+import framer from "vite-plugin-framer";
 
 // https://vitejs.dev/config/
 export default defineConfig({
-    plugins: [react(), mkcert(), framer()],
-    build: {
-        target: "ES2022",
-    },
-})
+	plugins: [react(), mkcert(), framer()],
+	build: {
+		target: "ES2022",
+	},
+});
